Power: Low Battery state fixes
Debugging low battery detection using an
adjustable power supply showed to issues

* Integer overflow computing milliseconds for 10min duration
  Math was done using 16bit, needed to add ul suffix to force 32bit

* Reassignment to a reference variable is not working. Use pointer
  instead

Strange that no compiler warnings where generated by this issues.

+++ src/hal/Timer/WakeUpTimer.h
@@ -51,7 +51,7 @@ namespace hal
              * This means a 8 second sleep at prescale 1024 and 8 bit count.
              * 
              */
-            static const uint32_t WAKEUP_INTERVAl_MS = 8000u;
+            static const uint16_t WAKEUP_INTERVAl_MS = 8000u;
 
             /** low level init of timer (does not enable it)
              */
